mcp_client_server/mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "google>=3.0.0",
    "google-genai>=1.25.0",
    "mcp>=1.11.0",
    "python-dotenv>=1.1.1",
]
